requirements.txt
aes==1.0.0
atomicwrites==1.4.0
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
crypto==1.4.1
decorator==4.4.2
future==0.18.2
huepy==1.2.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
iniconfig==1.0.1
instabot==0.117.0
mock==4.0.2
more-itertools==8.5.0
moviepy==1.0.3
Naked==0.1.31
numpy==1.19.2
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
proglog==0.1.9
py==1.9.0
pyasn1==0.4.8
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.0.2
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-toolbelt==0.9.1
responses==0.12.0
rsa==4.6
schedule==0.6.0
shellescape==3.8.1
six==1.15.0
toml==0.10.1
tqdm==4.49.0
urllib3==1.25.10
